MaliciousAppDetector/requirements.txt
fastapi
uvicorn
requests
scikit-learn
numpy
python-dotenv
